requirements.txt
flask
django
biopython
matplotlib
plotly
seaborn
pandas
numpy
scipy
